+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 * Fixed return more than one row error if real error raised on try read next row
 * Fixed checking errors for session must be deleted
+* Changed signature of filter func in balancers (replaced argument from `conn.Conn` type to `endpoint.Info`)
 
 ## v3.75.0
 * Improve config validation before start topic reader